phonenumber-service/src/main/java/com/jumia/phonenumbers/services/RegexValidator.java
package com.jumia.phonenumbers.services;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.regex.Matcher;
import java.util.regex.Pattern;

@Service
@RequiredArgsConstructor
public class RegexValidator {

    public boolean validateText(String regex, String text) {
        Pattern pattern = Pattern.compile(regex);
        Matcher matcher = pattern.matcher(text);
        while (matcher.find()) {
            return true;
        }
        return false;
    }

}
